Felix Update
- installed unity 2023.1.12f in order to use the localization package
- you may need to reinstall the cinemachine package
- you may need to reinstall the input system package
- you will likely need to reimport all materials

I've pushed every cahnge this time as I will need to be sure to push the localization settings too and not leave any changes due to updated unity version lying behind. lmk if you encounter issues with a newer version.

IMPORTANT: DO NOT!!! push back an older version of Unity as it will kick out the localization